requirements.txt
# Core
llama-index>=0.11.0
llama-index-llms-ollama>=0.3.0
llama-index-embeddings-ollama>=0.3.0
llama-index-vector-stores-chroma>=0.2.0
chromadb>=0.5.5

# Doc loaders
pypdf>=4.2.0
python-docx>=1.1.2
openpyxl>=3.1.5

# Web UI
streamlit>=1.39.0

# Utils
pyyaml>=6.0.2
